requirements.txt
django
wikipedia
nltk
djangorestframework
torch
torchvision